+++ src/Views/Filters/DateRangeFilter.php
@@ -183,26 +183,13 @@ class DateRangeFilter extends Filter
             }
 
             return $startDate.' to '.$endDate;
-
         }
 
         return '';
-
     }
 
-    public function render(string $filterLayout, string $tableName, bool $isTailwind, bool $isBootstrap4, bool $isBootstrap5): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
+    public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        $this->getOptions();
-        $this->getConfigs();
-
-        return view('livewire-tables::components.tools.filters.date-range', [
-            'filterLayout' => $filterLayout,
-            'tableName' => $tableName,
-            'isTailwind' => $isTailwind,
-            'isBootstrap' => ($isBootstrap4 || $isBootstrap5),
-            'isBootstrap4' => $isBootstrap4,
-            'isBootstrap5' => $isBootstrap5,
-            'filter' => $this,
-        ]);
+        return view('livewire-tables::components.tools.filters.date-range', $this->getFilterDisplayData());
     }
 }